Refactor tests, and fix help messages
